requirements.txt
asgiref==3.3.1
Django==3.1.4
django-model-utils==4.1.1
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
Unipath==1.1
